Make separate visitor for traversing rustc ty structs.
